Make all MultiMap works the same way
