+++ src/main/java/io/vertx/core/http/impl/HeadersAdaptor.java
@@ -19,6 +19,8 @@ package io.vertx.core.http.impl;
 import io.netty.handler.codec.http.HttpHeaders;
 import io.vertx.core.MultiMap;
 
+import java.util.ArrayList;
+import java.util.Collections;
 import java.util.Iterator;
 import java.util.List;
 import java.util.Map;
@@ -46,7 +48,14 @@ public class HeadersAdaptor implements MultiMap {
 
   @Override
   public List<Map.Entry<String, String>> entries() {
-    return headers.entries();
+    if (isEmpty()) {
+      return Collections.emptyList();
+    }
+    List<Map.Entry<String, String>> list = new ArrayList<Map.Entry<String, String>>(headers.size());
+    for (Map.Entry<String, String> entry : this) {
+      list.add(entry);
+    }
+    return list;
   }
 
   @Override
@@ -127,7 +136,44 @@ public class HeadersAdaptor implements MultiMap {
 
   @Override
   public Iterator<Map.Entry<String, String>> iterator() {
-    return headers.iterator();
+    return new Iterator<Map.Entry<String, String>>() {
+      final Iterator<Map.Entry<String, String>> iterator;
+
+      {
+        this.iterator = headers.iterator();
+      }
+
+      @Override
+      public boolean hasNext() {
+        return iterator.hasNext();
+      }
+
+      @Override
+      public Map.Entry<String, String> next() {
+        Map.Entry<String, String> next = iterator.next();
+        return new Map.Entry<String, String>() {
+          @Override
+          public String getKey() {
+            return next.getKey();
+          }
+
+          @Override
+          public String getValue() {
+            return next.getValue();
+          }
+
+          @Override
+          public String setValue(String value) {
+            return next.setValue(value);
+          }
+
+          @Override
+          public String toString() {
+            return next.getKey() + ": " + next.getValue();
+          }
+        };
+      }
+    };
   }
 
   @Override
@@ -186,5 +232,14 @@ public class HeadersAdaptor implements MultiMap {
   public MultiMap remove(CharSequence name) {
     headers.remove(name);
     return this;
+  }
+
+  @Override
+  public String toString() {
+    StringBuilder sb = new StringBuilder();
+    for (Map.Entry<String, String> entry: this) {
+      sb.append(entry).append('\n');
+    }
+    return sb.toString();
   }
 }
